refactor(ts/components/follower): move follower and recenterControl out of map.tsx
