+++ assets/tests/components/map.test.tsx
@@ -15,11 +15,11 @@ import React, { MutableRefObject } from "react"
 import { act } from "@testing-library/react"
 import { Map as LeafletMap } from "leaflet"
 import Map, {
-  autoCenter,
   defaultCenter,
   MapFollowingPrimaryVehicles,
   MapFollowingSelectionKey,
 } from "../../src/components/map"
+import { autoCenter } from "../../src/components/map/follower"
 import { TrainVehicle, VehicleInScheduledService } from "../../src/realtime"
 import vehicleFactory from "../factories/vehicle"
 import stopFactory from "../factories/stop"
